requirements.txt
flask==3.0.0
groq==0.4.2
python-dotenv==1.0.0
Pillow==11.0.0
openpyxl==3.1.2
pandas==2.2.0
PyPDF2==3.0.1
python-docx==1.0.1
